[hotfix] Rename QS ITCases to end in ITCase.
