+++ flink-queryable-state/flink-queryable-state-java/src/test/java/org/apache/flink/queryablestate/itcases/NonHAQueryableStateRocksDBBackendITCase.java
@@ -18,22 +18,28 @@
 
 package org.apache.flink.queryablestate.itcases;
 
+import org.apache.flink.contrib.streaming.state.RocksDBStateBackend;
 import org.apache.flink.runtime.state.AbstractStateBackend;
-import org.apache.flink.runtime.state.filesystem.FsStateBackend;
 
+import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.rules.TemporaryFolder;
 
 /**
- * Several integration tests for queryable state using the {@link FsStateBackend}.
+ * Several integration tests for queryable state using the {@link RocksDBStateBackend}.
  */
-public class HAQueryableStateITCaseFsBackend extends HAAbstractQueryableStateITCase {
+public class NonHAQueryableStateRocksDBBackendITCase extends NonHAAbstractQueryableStateTestBase {
 
 	@Rule
 	public TemporaryFolder temporaryFolder = new TemporaryFolder();
 
+	@BeforeClass
+	public static void setup() {
+		setup(9094, 9099);
+	}
+
 	@Override
 	protected AbstractStateBackend createStateBackend() throws Exception {
-		return new FsStateBackend(temporaryFolder.newFolder().toURI().toString());
+		return new RocksDBStateBackend(temporaryFolder.newFolder().toURI().toString());
 	}
 }
